+++ awscli/customizations/awslambda.py
@@ -10,13 +10,29 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import zipfile
+from contextlib import closing
+
+from botocore.vendored import six
+
 from awscli.arguments import CustomArgument
 from awscli.customizations import utils
+
+ERROR_MSG = (
+    "--zip-file must be a file with the fileb:// prefix.\n"
+    "Example usage:  --zip-file fileb://path/to/file.zip")
 
 
 def register_lambda_create_function(cli):
     cli.register('building-argument-table.lambda.create-function',
                  _flatten_code_argument)
+    cli.register('process-cli-arg.lambda.update-function-code',
+                 validate_is_zip_file)
+
+
+def validate_is_zip_file(cli_argument, value, **kwargs):
+    if cli_argument.name == 'zip-file':
+        _should_contain_zip_content(value)
 
 
 def _flatten_code_argument(argument_table, **kwargs):
@@ -27,7 +43,25 @@ def _flatten_code_argument(argument_table, **kwargs):
     del argument_table['code']
 
 
+def _should_contain_zip_content(value):
+    if not isinstance(value, bytes):
+        # If it's not bytes it's basically impossible for
+        # this to be valid zip content, but we'll at least
+        # still try to load the contents as a zip file
+        # to be absolutely sure.
+        value = value.encode('utf-8')
+    fileobj = six.BytesIO(value)
+    try:
+        with closing(zipfile.ZipFile(fileobj)) as f:
+            f.infolist()
+    except zipfile.BadZipfile:
+        raise ValueError(ERROR_MSG)
+
+
 class ZipFileArgument(CustomArgument):
     def add_to_params(self, parameters, value):
+        if value is None:
+            return
+        _should_contain_zip_content(value)
         zip_file_param = {'ZipFile': value}
         parameters['Code'] = zip_file_param