+++ awscli/paramfile.py
@@ -11,7 +11,6 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-
 import logging
 import os
 
@@ -74,32 +73,43 @@ class ResourceLoadingError(Exception):
 
 
 def get_paramfile(path):
-    """
+    """Load parameter based on a resource URI.
+
     It is possible to pass parameters to operations by referring
     to files or URI's.  If such a reference is detected, this
     function attempts to retrieve the data from the file or URI
     and returns it.  If there are any errors or if the ``path``
     does not appear to refer to a file or URI, a ``None`` is
     returned.
+
+    :type path: str
+    :param path: The resource URI, e.g. file://foo.txt.  This value
+        may also be a non resource URI, in which case ``None`` is returned.
+
+    :return: The loaded value associated with the resource URI.
+        If the provided ``path`` is not a resource URI, then a
+        value of ``None`` is returned.
+
     """
     data = None
     if isinstance(path, six.string_types):
-        for prefix in PrefixMap:
+        for prefix, function_spec in PREFIX_MAP.items():
             if path.startswith(prefix):
-                kwargs = KwargsMap.get(prefix, {})
-                data = PrefixMap[prefix](prefix, path, **kwargs)
+                function, kwargs = function_spec
+                data = function(prefix, path, **kwargs)
     return data
 
 
 def get_file(prefix, path, mode):
-    file_path = path[len(prefix):]
-    file_path = os.path.expanduser(file_path)
-    file_path = os.path.expandvars(file_path)
-    if not os.path.isfile(file_path):
-        raise ResourceLoadingError("file does not exist: %s" % file_path)
+    file_path = os.path.expandvars(os.path.expanduser(path[len(prefix):]))
     try:
         with compat_open(file_path, mode) as f:
             return f.read()
+    except UnicodeDecodeError:
+        raise ResourceLoadingError(
+            'Unable to load paramfile (%s), text contents could '
+            'not be decoded.  If this is a binary file, please use the '
+            'fileb:// prefix instead of the file:// prefix.' % file_path)
     except (OSError, IOError) as e:
         raise ResourceLoadingError('Unable to load paramfile %s: %s' % (
             path, e))
@@ -118,12 +128,9 @@ def get_uri(prefix, uri):
         raise ResourceLoadingError('Unable to retrieve %s: %s' % (uri, e))
 
 
-PrefixMap = {'file://': get_file,
-             'fileb://': get_file,
-             'http://': get_uri,
-             'https://': get_uri}
-
-KwargsMap = {'file://': {'mode': 'r'},
-             'fileb://': {'mode': 'rb'},
-             'http://': {},
-             'https://': {}}
+PREFIX_MAP = {
+    'file://': (get_file, {'mode': 'r'}),
+    'fileb://': (get_file, {'mode': 'rb'}),
+    'http://': (get_uri, {}),
+    'https://': (get_uri, {}),
+}