+++ tests/unit/customizations/test_awslambda.py
@@ -10,52 +10,48 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+import os
+import zipfile
+from contextlib import closing
+
 from awscli.testutils import unittest
 from awscli.testutils import BaseAWSCommandParamsTest
 from awscli.testutils import FileCreator
 
 
-class TestCreateFunction(BaseAWSCommandParamsTest):
-
-    prefix = 'lambda create-function'
+class BaseLambdaTests(BaseAWSCommandParamsTest):
 
     def setUp(self):
-        super(TestCreateFunction, self).setUp()
-
-        # Make a temporary file
+        super(BaseLambdaTests, self).setUp()
         self.files = FileCreator()
-        self.contents_of_file = 'myzipcontents'
         self.temp_file = self.files.create_file(
-            'foo', self.contents_of_file)
+            'foo', 'mycontents')
+        self.zip_file = os.path.join(self.files.rootdir, 'foo.zip')
+        with closing(zipfile.ZipFile(self.zip_file, 'w')) as f:
+            f.write(self.temp_file)
+        with open(self.zip_file, 'rb') as f:
+            self.zip_file_contents = f.read()
 
     def tearDown(self):
-        super(TestCreateFunction, self).tearDown()
+        super(BaseLambdaTests, self).tearDown()
         self.files.remove_all()
 
-    def test_create_function(self):
-        cmdline = self.prefix
-        cmdline += ' --function-name myfunction --runtime myruntime'
-        cmdline += ' --role myrole --handler myhandler --zip-file myzip'
-        result = {
-            'FunctionName': 'myfunction',
-            'Runtime': 'myruntime',
-            'Role': 'myrole',
-            'Handler': 'myhandler',
-            'Code': {'ZipFile': 'myzip'}
-        }
-        self.assert_params_for_cmd(cmdline, result)
+
+class TestCreateFunction(BaseLambdaTests):
+
+    prefix = 'lambda create-function'
 
     def test_create_function_with_file(self):
         cmdline = self.prefix
         cmdline += ' --function-name myfunction --runtime myruntime'
         cmdline += ' --role myrole --handler myhandler'
-        cmdline += ' --zip-file file://%s' % self.temp_file
+        cmdline += ' --zip-file fileb://%s' % self.zip_file
         result = {
             'FunctionName': 'myfunction',
             'Runtime': 'myruntime',
             'Role': 'myrole',
             'Handler': 'myhandler',
-            'Code': {'ZipFile': self.contents_of_file}
+            'Code': {'ZipFile': self.zip_file_contents}
         }
         self.assert_params_for_cmd(cmdline, result)
 
@@ -66,3 +62,48 @@ class TestCreateFunction(BaseAWSCommandParamsTest):
         cmdline += ' --code mycode'
         stdout, stderr, rc = self.run_cmd(cmdline, expected_rc=255)
         self.assertIn('Unknown options: --code', stderr)
+
+    def test_create_function_with_invalid_file_contents(self):
+        cmdline = self.prefix
+        cmdline += ' --function-name myfunction --runtime myruntime'
+        cmdline += ' --role myrole --handler myhandler'
+        cmdline += ' --zip-file filename_instead_of_contents.zip'
+        stdout, stderr, rc = self.run_cmd(cmdline, expected_rc=255)
+        self.assertIn('must be a file with the fileb:// prefix', stderr)
+        # Should also give a pointer to fileb:// for them.
+        self.assertIn('fileb://', stderr)
+
+    def test_not_using_fileb_prefix(self):
+        cmdline = self.prefix
+        cmdline += ' --function-name myfunction --runtime myruntime'
+        cmdline += ' --role myrole --handler myhandler'
+        # Note file:// instead of fileb://
+        cmdline += ' --zip-file file://%s' % self.zip_file
+        stdout, stderr, rc = self.run_cmd(cmdline, expected_rc=255)
+        # Ensure we mention fileb:// to give the user an idea of
+        # where to go next.
+        self.assertIn('fileb://', stderr)
+
+
+class TestUpdateFunctionCode(BaseLambdaTests):
+
+    prefix = 'lambda update-function-code'
+
+    def test_not_using_fileb_prefix(self):
+        cmdline = self.prefix + ' --function-name foo'
+        cmdline += ' --zip-file filename_instead_of_contents.zip'
+        stdout, stderr, rc = self.run_cmd(cmdline, expected_rc=255)
+        self.assertIn('must be a file with the fileb:// prefix', stderr)
+        # Should also give a pointer to fileb:// for them.
+        self.assertIn('fileb://', stderr)
+
+    def test_using_fileb_prefix_succeeds(self):
+        cmdline = self.prefix
+        cmdline += ' --function-name myfunction'
+        cmdline += ' --zip-file fileb://%s' % self.zip_file
+        result = {
+            'FunctionName': 'myfunction',
+            'ZipFile': self.zip_file_contents,
+        }
+        self.assert_params_for_cmd(cmdline, result)
+